reorganize files and create small room card comp
